orchestrator/requirements.txt
# FastAPI & Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Pipecat for Voice Pipeline
# daily: Daily.co transport support
# webrtc: SmallWebRTC for local mode
# silero: VAD (Voice Activity Detection)
pipecat-ai[daily,webrtc,silero]==0.0.99
opencv-python-headless
onnxruntime

# RAG Components
qdrant-client>=1.11.0,<1.13.0
sentence-transformers>=2.2.2

# Settings & Utils
pydantic-settings>=2.1.0
httpx>=0.26.0
aiohttp>=3.9.0

# Audio Processing
numpy>=1.24.0

# WebRTC
aiortc>=1.6.0
aioice>=0.9.0
